Extract RelatedInverseLinksProcessor from cache
